requirements.txt
streamlit==1.45.0
google-generativeai==0.8.3
faiss-cpu
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
sentence-transformers==3.4.1
plotly==5.22.0
pandas==2.2.3
numpy==2.2.3